requirements.txt
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
bleach==3.1.5
boto3==1.15.16
botocore==1.18.16
Brotli==1.0.7
cachetools==4.1.1
certifi==2019.9.11
chardet==3.0.4
chatbot==1.5.2b0
click==7.1.2
colorama==0.4.3
cycler==0.10.0
dash==1.14.0
dash-core-components==1.10.2
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.42.0
django-crispy-forms==1.9.1
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.10.1
djangorestframework==3.11.0
docopt==0.6.2
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
google-api-core==1.21.0
google-api-python-client==1.9.3
google-auth==1.18.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.8
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
Js2Py==0.66
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.4
jupyterlab-server==1.1.5
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.6
nltk==3.4.5
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
opencv-python==4.2.0.34
packaging==19.2
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.1
Pillow-PIL==0.1.dev0
pipwin==0.4.5
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.2
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyjsparser==2.7.1
pylint==2.5.3
pyparsing==2.4.4
PyPrind==2.11.2
pyrsistent==0.16.0
python-dateutil==2.8.1
python3-openid==3.2.0
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
robobrowser==0.5.3
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
tzlocal==2.0.0
uritemplate==3.0.1
urllib3==1.25.6
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==0.16.0
whitenoise==5.2.0
wrapt==1.12.1
